Move bma421 enum to a public header
