+++ drivers/sensor/bma421/bma421.h
@@ -12,22 +12,10 @@
 #include <zephyr/types.h>
 #include <drivers/gpio.h>
 #include <drivers/sensor.h>
-
+#include <drivers/sensor/bma421.h>
 #include "bma4_defs.h"
 
 #define BMA421_I2C_ADDRESS		DT_INST_REG_ADDR(0)
-
-enum bma421_channel {
-	BMA421_CHAN_STEP_COUNTER = SENSOR_CHAN_PRIV_START,
-	BMA421_CHAN_STEP_ENABLE,
-};
-
-/* If necessary at some point */
-enum bma421_attribute {
-	BMA421_ATTR_CPI = SENSOR_ATTR_PRIV_START,
-	BMA421_ATTR_REST_ENABLE,
-	BMA421_ATTR_RUN_DOWNSHIFT_TIME,
-};
 
 struct bma421_data {
 	struct device *i2c;
